Tests: Try to discover write-lock issue on Windows
Actually, the `test_upload_files_success` test should have revealed it.
However, we discovered that when mocking the `NamedTemporaryFile`
object, we created an instance instead of propagating a factory.

That is wrong, because it expands the scope of the temporary file to
the whole test case, and not the scope of the `test_upload_files`
function only.